requirements.txt
Pillow==4.0.0
django==1.10.5
djangorestframework==3.5.3
pytest==3.0.5
pytest-cov==2.4.0
Faker==0.7.7
mock==2.0.0
psycopg2==2.6.2
pytest-django==3.1.2